feat: describe values
Creates a function that generates a complete description of the exported
fields of an arbitrary value recursively. There is both a free function
and a method on a Vistor type. The free function will always use a
default Visitor type and the Visitor method will use configurable
properties to change the description behaviour.